bytesize

Struct ByteSize

pub struct ByteSize(pub u64);

Byte size representation.

Tuple Fields
0: u64
Implementations

impl ByteSize

pub const fn b(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of bytes.

pub const fn kb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of kilobytes.

pub const fn kib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of kibibytes.

pub const fn mb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of megabytes.

pub const fn mib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of mebibytes.

pub const fn gb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of gigabytes.

pub const fn gib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of gibibytes.

pub const fn tb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of terabytes.

pub const fn tib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of tebibytes.

pub const fn pb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of petabytes.

pub const fn pib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of pebibytes.

pub const fn eb(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of exabytes.

pub const fn eib(size: u64) -> ByteSize

Constructs a byte size wrapper from a quantity of exbibytes.

pub const fn as_u64(&self) -> u64

Returns byte count.

pub fn as_kb(&self) -> f64

Returns byte count as kilobytes.

pub fn as_kib(&self) -> f64

Returns byte count as kibibytes.

pub fn as_mb(&self) -> f64

Returns byte count as megabytes.

pub fn as_mib(&self) -> f64

Returns byte count as mebibytes.

pub fn as_gb(&self) -> f64

Returns byte count as gigabytes.

pub fn as_gib(&self) -> f64

Returns byte count as gibibytes.

pub fn as_tb(&self) -> f64

Returns byte count as terabytes.

pub fn as_tib(&self) -> f64

Returns byte count as tebibytes.

pub fn as_pb(&self) -> f64

Returns byte count as petabytes.

pub fn as_pib(&self) -> f64

Returns byte count as pebibytes.

pub fn as_eb(&self) -> f64

Returns byte count as exabytes.

pub fn as_eib(&self) -> f64

Returns byte count as exbibytes.

pub fn display(&self) -> Display

Returns a formatting display wrapper.